requirements.txt
streamlit
langchain-core
langchain-openai
python-dotenv
